refactor: ShowDetail 렌더링 위치를 VenueDetail에서 ShowList로 변경
