+++ fe/user/src/pages/MapPage/Panel/VenueDetail/RestInfo/ShowInfo/ShowList.tsx
@@ -1,15 +1,16 @@
 import styled from '@emotion/styled';
 import ChevronLeftIcon from '@mui/icons-material/ChevronLeft';
 import ChevronRightIcon from '@mui/icons-material/ChevronRight';
-import { ShowDetail } from '~/types/api.types';
+import { useState } from 'react';
+import { ShowDetail as ShowDetailData } from '~/types/api.types';
 import { getKoreanWeekdayName } from '~/utils/dateUtils';
+import { ShowDetail } from '../../ShowDetail';
 
 type Props = {
-  showList?: ShowDetail[];
+  showList?: ShowDetailData[];
   selectedDate: Date;
   selectPreviousDate: () => void;
   selectNextDate: () => void;
-  onShowListClick: (shows: ShowDetail[], showIndex: number) => void;
 };
 
 export const ShowList: React.FC<Props> = ({
@@ -17,12 +18,21 @@ export const ShowList: React.FC<Props> = ({
   selectedDate,
   selectPreviousDate,
   selectNextDate,
-  onShowListClick,
 }) => {
+  const [currentShowIndex, setCurrentShowIndex] = useState(-1);
+
   const month = selectedDate.getMonth() + 1;
   const date = selectedDate.getDate();
   const weekName = getKoreanWeekdayName(selectedDate.getDay());
   const dateString = `${month}월 ${date}일 ${weekName}요일`;
+
+  const openShowDetail = (index: number) => {
+    setCurrentShowIndex(index);
+  };
+
+  const hideShowDetail = () => {
+    setCurrentShowIndex(-1);
+  };
 
   return (
     <StyledShowList>
@@ -44,7 +54,7 @@ export const ShowList: React.FC<Props> = ({
           showList.map((show, index) => (
             <StyledShowListItem
               key={show.id}
-              onClick={() => onShowListClick(showList, index)}
+              onClick={() => openShowDetail(index)}
             >
               <StyledShowListItemIndex>
                 {String(index + 1).padStart(2, '0')}
@@ -58,6 +68,13 @@ export const ShowList: React.FC<Props> = ({
               </StyledShowListItemTime>
             </StyledShowListItem>
           ))}
+        {showList && currentShowIndex !== -1 && (
+          <ShowDetail
+            showList={showList}
+            currentIndex={currentShowIndex}
+            onCloseClick={hideShowDetail}
+          />
+        )}
       </StyledShowListContent>
     </StyledShowList>
   );
