+++ fe/user/src/pages/MapPage/Panel/VenueDetail/index.tsx
@@ -10,15 +10,11 @@ import {
 } from 'react-router-dom';
 import { getVenueDetail } from '~/apis/venue';
 import CaretLeft from '~/assets/icons/CaretLeft.svg?react';
-import {
-  ShowDetail as ShowDetailData,
-  VenueDetailData,
-} from '~/types/api.types';
+import { VenueDetailData } from '~/types/api.types';
 import { BasicInfo } from './BasicInfo';
 import { Header } from './Header';
 import { Images } from './Images';
 import { RestInfo } from './RestInfo';
-import { ShowDetail } from './ShowDetail';
 
 export const VenueDetail: React.FC = () => {
   const { venueId } = useParams();
@@ -27,21 +23,9 @@ export const VenueDetail: React.FC = () => {
   const mapElement = useOutletContext<React.RefObject<HTMLDivElement>>();
   const [isRender, setRender] = useState(false);
   const [data, setData] = useState<VenueDetailData>();
-  const [showList, setShowList] = useState<ShowDetailData[]>([]);
-  const [currentShowIndex, setCurrentShowIndex] = useState(-1);
 
   const backToVenueList = () => {
     navigate(`/map?${currentLocation.search}`);
-  };
-
-  const openShowDetail = (shows: ShowDetailData[], showIndex: number) => {
-    setShowList(shows);
-    setCurrentShowIndex(showIndex);
-  };
-
-  const hideShowDetail = () => {
-    setShowList([]);
-    setCurrentShowIndex(-1);
   };
 
   useEffect(() => {
@@ -88,18 +72,8 @@ export const VenueDetail: React.FC = () => {
               data.links.find((link) => link.type === 'naverMap')?.url
             }
           />
-          <RestInfo
-            description={data.description}
-            onShowListClick={openShowDetail}
-          />
+          <RestInfo description={data.description} />
           <Outlet context={mapElement} />
-          {showList.length > 0 && currentShowIndex !== -1 && (
-            <ShowDetail
-              showList={showList}
-              currentIndex={currentShowIndex}
-              onCloseClick={hideShowDetail}
-            />
-          )}
         </StyledVenueDetail>
       )}
     </>
